# Conflicts:
#	crates/cargo-fe2o3/src/authorized_kernel_closure.rs
#	crates/fe2o3-amdgcn-model/src/production_kir_to_llvm_replay_v1.rs
#	crates/fe2o3-amdgcn-model/src/production_refinement_v1.rs
#	crates/fe2o3-compiler-execution-coordinator/README.md
#	crates/fe2o3-compiler-execution-coordinator/src/provisioning_entrypoint.rs
#	crates/fe2o3-compiler-lineage/src/lib.rs
#	crates/fe2o3-hsaco-finalize/tests/worker_v3_hsaco_admission.rs
#	crates/fe2o3-lower-mir-kernel/src/production_semantic_kir_v1.rs
#	crates/fe2o3-mir-model/src/semantic_mir_v1/canonical_decode.rs
#	crates/rustc-codegen-fe2o3/src/production_pipeline.rs
#	crates/rustc-codegen-fe2o3/src/production_ranked_projection_v1.rs
#	crates/rustc-codegen-fe2o3/src/production_semantic_body_v1.rs
#	crates/rustc-codegen-fe2o3/src/production_semantic_lineage_v3.rs
#	crates/rustc-codegen-fe2o3/src/trusted_device_items.rs
#	deployment/README.md
#	scripts/tests/compiler-execution-systemd.sh

+++ crates/fe2o3-artifact-transaction/src/publication_object_inventory_v1.rs
@@ -175,7 +175,7 @@ impl PublicationObjectRootIdentityV1 {
             stat.st_uid,
             stat.st_gid,
             stat.st_mode,
-            stat.st_nlink as u64,
+            stat.st_nlink,
         )
     }
 
@@ -185,7 +185,7 @@ impl PublicationObjectRootIdentityV1 {
             && self.owner_uid == stat.st_uid
             && self.owner_gid == stat.st_gid
             && self.mode == stat.st_mode
-            && self.link_count == stat.st_nlink as u64
+            && self.link_count == stat.st_nlink
     }
 }
